+++ rearq/cli.py
@@ -16,7 +16,10 @@ def coro(f):
     @wraps(f)
     def wrapper(*args, **kwargs):
         loop = asyncio.get_event_loop()
-        return loop.run_until_complete(f(*args, **kwargs))
+        try:
+            return loop.run_until_complete(f(*args, **kwargs))
+        except asyncio.CancelledError:
+            pass
 
     return wrapper